requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
npm==0.1.1
optional-django==0.1.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
yarn==0.0.9
